# openapi.yaml
openapi: 3.0.3
info:
  title: tr-engine API
  version: 0.9.1
  contact:
    name: tr-engine
    url: https://github.com/LumenPrima/tr-engine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Backend service for trunk-recorder data ingestion and querying.
    Provides REST APIs for accessing radio system data, calls, talkgroups,
    units, and recorder state.

    ## Authentication

    tr-engine supports three auth modes derived from configuration:

    - **Open mode**: no `AUTH_TOKEN` and no `ADMIN_PASSWORD`; requests do not
      require authentication.
    - **Token mode**: `AUTH_TOKEN` set and `ADMIN_PASSWORD` unset; clients use
      `Authorization: Bearer <token>` or `?token=`.
    - **Full mode**: `ADMIN_PASSWORD` set; browser users authenticate with JWT
      sessions and service clients can use `tre_...` API keys. If `AUTH_TOKEN`
      is also set, it acts as a public read token returned by `/auth-init`.

    Mutating operations (POST, PATCH, PUT, DELETE) require editor/admin JWT
    role, a suitable API key, or the deprecated legacy `WRITE_TOKEN`.

    ## Data Model: Systems and Sites

    The data model separates **logical radio networks** from **recording
    sites**:

    - **System** (`system_id`): a logical radio network. For P25, this
      is the WACN-level network identified by `(sysid, wacn)`. For
      conventional, it is identified by `(instance_id, sys_name)`.
...

...
      specific site.

    Example:
    ```
...
    ```

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/setup:
    get:
      operationId: checkAuthSetup
      summary: Check whether first-run setup is needed
      description: |
        Returns whether the instance has completed first-run admin setup.
        Unauthenticated — always available when JWT auth is configured.
        The dashboard calls this endpoint to decide whether to show the
        admin setup flow or the regular login page.
      tags: [auth]
      security: []
      responses:
        "200":
          description: Setup status
          content:
            application/json:
              schema:
                type: object
                properties:
                  needs_setup:
                    type: boolean
                    description: true when zero users exist (first-run setup pending)
                  user_count:
                    type: integer
                    description: Total number of users in the database
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      operationId: authSetup
      summary: First-run admin creation
      description: |
        Creates the first admin user. Only works when zero users exist in the
        database. Rate-limited to 5 attempts per minute per IP. Only available
        when JWT auth is configured (JWT_SECRET or ADMIN_PASSWORD set).

        Uses an atomic INSERT to prevent race conditions where two concurrent
        requests could both create admin users.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  description: Admin username (normalized to lowercase)
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: Admin user created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: "#/components/schemas/UserSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Setup already completed — users exist
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/login:
    post:
      operationId: authLogin
      summary: User login
      description: |
        Validates credentials and returns a JWT access token. Also sets an
        httpOnly refresh cookie (`tr_refresh_token`). Rate-limited to 5
        attempts per minute per IP.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token (1 hour expiry)
                  user:
                    $ref: "#/components/schemas/UserSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials or account disabled
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/refresh:
    post:
      operationId: authRefresh
      summary: Refresh access token
      description: |
        Validates the httpOnly refresh cookie and returns a new JWT access token.
        The refresh token has a 7-day expiry.
      tags: [auth]
      security: []
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  user:
                    $ref: "#/components/schemas/UserSummary"
        "401":
          description: No refresh token or invalid/expired token
...
  /auth/me:
    get:
      operationId: authMe
      summary: Current user info
      description: Returns the authenticated user's profile from the database.
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /auth/keys:
    get:
      operationId: listOwnKeys
      summary: List own API keys
      description: Returns API keys owned by the authenticated user. Requires editor or admin role.
      tags: [auth]
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKey"
        "401":
          description: User authentication required (API keys cannot list keys)
    post:
      operationId: createKey
      summary: Create API key
      description: |
        Creates a new API key. The plaintext key is returned once in the response
        and cannot be retrieved again. The key's role is capped at the caller's
        own role level — an editor cannot create an admin key.
        Requires editor or admin role.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:
                  type: string
                  description: Human-readable name for the key
                role:
                  type: string
                  enum: [viewer, editor, admin]
                  default: viewer
      responses:
        "201":
          description: API key created (plaintext shown once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyWithPlaintext"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Cannot create key with higher role than your own
  /auth/keys/{id}:
    delete:
      operationId: deleteOwnKey
      summary: Revoke own API key
      description: Deletes an API key owned by the authenticated user. Requires editor or admin role.
      tags: [auth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Key revoked
        "404":
          description: Key not found or not owned by you
  /auth/keys/all:
    get:
      operationId: listAllKeys
      summary: List all API keys
      description: Returns all API keys across all users and service accounts. Admin only.
      tags: [auth]
      responses:
        "200":
          description: List of all API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKey"
        "403":
          description: Admin access required
  /auth/keys/service:
    post:
      operationId: createServiceAccountKey
      summary: Create service account key
      description: |
        Creates an API key that acts as its own identity, not linked to any user.
        Used for frontend apps, bots, and integrations. Admin only.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:
                  type: string
                role:
                  type: string
                  enum: [viewer, editor, admin]
                  default: viewer
      responses:
        "201":
          description: Service account key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyWithPlaintext"
        "403":
          description: Admin access required
  /auth/keys/{id}/any:
    delete:
      operationId: deleteAnyKey
      summary: Revoke any API key
      description: Deletes any API key by ID regardless of ownership. Admin only.
      tags: [auth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Key revoked
        "403":
          description: Admin access required
        "404":
          description: Key not found
  /users:
    get:
      operationId: listUsers
      summary: List all users
      description: Returns all user accounts. Admin only.
      tags: [auth]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  total:
                    type: integer
        "403":
          description: Admin access required
    post:
      operationId: createUser
      summary: Create user
      description: Creates a new user account. Admin only.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  minLength: 8
                role:
                  type: string
                  enum: [viewer, editor, admin]
                  default: viewer
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Admin access required
        "409":
          description: Username already exists
  /users/{id}:
    patch:
      operationId: updateUser
      summary: Update user
      description: |
        Partially update a user account (role, password, enabled status).
        Admin only. Cannot demote or disable the last admin user.
        Cannot demote or disable your own account.
      tags: [auth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [viewer, editor, admin]
                password:
                  type: string
                  minLength: 8
                enabled:
                  type: boolean
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Admin access required, or last admin protection
        "404":
          description: User not found
    delete:
      operationId: deleteUser
      summary: Delete user
      description: |
        Deletes a user and cascades to their sessions and API keys.
        Admin only. Cannot delete yourself or the last admin user.
      tags: [auth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
        "403":
          description: Admin access required, or last admin / self-delete protection
        "404":
          description: User not found

  # ----------------------------------------------------------
  # Systems
  # ----------------------------------------------------------
...
  /pages:
    get:
      operationId: listPages
      summary: List web UI pages
      description: |
        Scans the web directory for HTML pages with `card-title` meta tags
        and returns metadata for building navigation. Pages without
        `card-title` are excluded.
      tags: [health]
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      description: Relative path to the HTML page
                      example: /events.html
                    title:
                      type: string
                      description: Page title from card-title meta tag
                      example: Live Events
                    description:
                      type: string
                      description: Page description from card-description meta tag
                      example: Real-time event stream
                    order:
                      type: integer
                      description: Sort order (lower = first)
                      example: 2
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: savePage
      summary: Save an HTML page to the web directory
      description: |
        Writes an HTML file to the server's disk `web/` directory, where
        the page discovery system picks it up immediately. Requires the
        HTML to contain a `card-title` meta tag. Protected core files
        (index.html, auth.js, theme-config.js, theme-engine.js,
        playground.html) cannot be overwritten. Returns 503 if the server
        is using embedded web files with no disk directory.
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, html]
              properties:
                filename:
                  type: string
                  description: Target filename (sanitized server-side; .html appended if missing)
                  example: my-dashboard.html
                html:
                  type: string
                  description: Full HTML content of the page (max 1 MB)
                overwrite:
                  type: boolean
                  default: false
                  description: If true, overwrite an existing file with the same name
      responses:
        "201":
          description: Page saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                    example: /my-dashboard.html
                  title:
                    type: string
                    example: My Dashboard
                  message:
                    type: string
                    example: Page saved
        "400":
          description: Invalid request (empty filename, missing card-title, HTML too large)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Attempted to overwrite a protected core file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: File already exists and overwrite was not set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Server is using embedded web files (no disk web/ directory)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # OpenAPI Spec
  # ----------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Pass token in the Authorization header:
        `Authorization: Bearer <token>`

  # ----------------------------------------------------------
  # Reusable Parameters
  # ----------------------------------------------------------
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Too many requests — rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [viewer, editor, admin]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [viewer, editor, admin]
        enabled:
          type: boolean
        display_name:
          type: string
        last_login:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    APIKey:
      type: object
      properties:
        id:
          type: integer
        key_prefix:
          type: string
          description: First 12 characters of the key for identification
        user_id:
          type: integer
          nullable: true
          description: Null for service accounts
        role:
          type: string
          enum: [viewer, editor, admin]
        label:
          type: string
        is_service_account:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
    APIKeyWithPlaintext:
      allOf:
        - $ref: "#/components/schemas/APIKey"
        - type: object
          properties:
            key:
              type: string
              description: |
                The full API key plaintext. Shown exactly once at creation —
                cannot be retrieved again. Store it securely.

    # ========== Error Models ==========
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: Machine-readable error code for programmatic handling
          example: not_found
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - service_unavailable
            - invalid_body
            - invalid_parameter
            - invalid_time_range
            - query_failed
            - ambiguous_id
            - duplicate
            - request_timeout
        error:
          type: string
          example: Resource not found
        detail:
          type: string
          description: Additional context when available
...